Add L2 Norm and Squared L2 Norm of Dimensions

These functions compute and add L2 norms and squared L2 norms of the frequency dimensions (.dim_* columns) as additional columns to a tidy_fft object.

Usage

add_l2nm(x)

add_l2sq(x)

Arguments

x

A tidy_fft object containing frequency dimensions (.dim_*) and associated metadata.

Value

A tidy_fft object with additional columns:

l2nm: The L2 norm of the frequency dimensions.

l2sq: The squared L2 norm of the frequency dimensions.

Details

add_l2nm(): Adds a column l2nm containing the L2 norm, calculated as the square root of the sum of squares across .dim_* columns.

add_l2sq(): Adds a column l2sq containing the squared L2 norm, calculated as the sum of squares across .dim_* columns.

See also

Examples

matrix(1:9, 3) |>
  tidy_fft() |>
  print(n = 3) |>
  add_l2nm() |>
  print(n = 3) |>
  add_l2sq() |>
  print(n = 3)
#> # A tibble: 9 × 3
#>   .dim_1 .dim_2 fx            
#> *  <dbl>  <dbl> <cpl>         
#> 1  0          0 45.0+0.000000i
#> 2  0.333      0 -4.5+2.598076i
#> 3 -0.333      0 -4.5-2.598076i
#> # ℹ 6 more rows
#> # A tibble: 9 × 4
#>   .dim_1 .dim_2 fx              l2nm
#>    <dbl>  <dbl> <cpl>          <dbl>
#> 1  0          0 45.0+0.000000i 0    
#> 2  0.333      0 -4.5+2.598076i 0.333
#> 3 -0.333      0 -4.5-2.598076i 0.333
#> # ℹ 6 more rows
#> # A tibble: 9 × 5
#>   .dim_1 .dim_2 fx              l2nm  l2sq
#>    <dbl>  <dbl> <cpl>          <dbl> <dbl>
#> 1  0          0 45.0+0.000000i 0     0    
#> 2  0.333      0 -4.5+2.598076i 0.333 0.111
#> 3 -0.333      0 -4.5-2.598076i 0.333 0.111
#> # ℹ 6 more rows