Check and Retrieve Representations of a tidy_fft Object

These functions check and retrieve specific representations of a tidy_fft object. Supported representations include:

Complex ("cplx"): Contains a column fx with complex Fourier coefficients.

Rectangular ("rect"): Contains columns re (real) and im (imaginary) components.

Polar ("polr"): Contains columns mod (modulus) and arg (argument).

Usage

can_repr(x, repr)

get_repr(x)

has_cplx(x)

has_rect(x)

has_polr(x)

Arguments

x

A tidy_fft object.

repr

For can_repr(), the target representation to check. A character vector with one or more of ("polr", "rect", or "cplx").

Value

can_repr(): A logical value (TRUE or FALSE) indicating if the object has any of the specified representations.

get_repr(): A character vector of representations present in the object.

has_cplx(), has_rect(), has_polr(): Logical values (TRUE or FALSE) indicating the presence of specific representations.

Examples

tft <- tidy_fft(c(1, 0, -1, 0))

# Check specific representations
can_repr(tft, "cplx") # TRUE
#> [1] TRUE
can_repr(tft, "rect") # FALSE
#> [1] FALSE

# Retrieve current representations
get_repr(tft) # "cplx"
#> [1] "cplx"

# Check individual representations
has_cplx(tft) # TRUE
#> [1] TRUE
has_rect(tft) # FALSE
#> [1] FALSE
has_polr(tft) # FALSE
#> [1] FALSE