Compute the Cross-Spectrum

This function computes the cross-spectrum between two tidy_fft objects by multiplying the Fourier coefficients (fx) of the first object (a) with the conjugate of the coefficients from the second object (b).

Usage

cross_spec(a, b)

Arguments

a

A tidy_fft object.

b

A tidy_fft object with the same number of rows as a.

Value

A tidy_fft object with updated fx values representing the cross-spectrum.

Details

The cross-spectrum is calculated as: $$S_{ab}(k) = X_a(k) \cdot \overline{X_b(k)}$$ where \(X_a(k)\) and \(X_b(k)\) are the Fourier coefficients of a and b, respectively, and \(\overline{X_b(k)}\) is the complex conjugate of \(X_b(k)\).

See also

Examples

# Generate example tidy_fft objects
a <- tidy_fft(c(1, 0, -1, 0))
b <- tidy_fft(c(0, 1, 0, -1))

# Compute the cross-spectrum
cross_result <- cross_spec(a, b)
print(cross_result)
#> # A tibble: 4 × 2
#>   dim_1 fx   
#>   <dbl> <cpl>
#> 1  0    0+0i 
#> 2  0.25 0+4i 
#> 3  0.5  0+0i 
#> 4 -0.25 0-4i