Compute Fourier frequencies for multidimensional arrays

Computes Fourier frequencies for each dimension of an array and returns a tidy tibble with frequencies for all dimensions.

Usage

# S3 method for class 'array'
fourier_frequencies(x)

Arguments

x

A numeric array.

Value

A tibble where each column corresponds to the Fourier frequencies of a dimension of the input array. The column names are dim_1, dim_2, ..., corresponding to each dimension.

See also

Examples

# Compute Fourier frequencies for a 3D array
array_input <- array(1:27, dim = c(3, 3, 3))
fourier_frequencies(array_input)
#> # A tibble: 27 × 3
#>    dim_3  dim_2  dim_1
#>    <dbl>  <dbl>  <dbl>
#>  1 0      0      0    
#>  2 0      0      0.333
#>  3 0      0     -0.333
#>  4 0      0.333  0    
#>  5 0      0.333  0.333
#>  6 0      0.333 -0.333
#>  7 0     -0.333  0    
#>  8 0     -0.333  0.333
#>  9 0     -0.333 -0.333
#> 10 0.333  0      0    
#> # ℹ 17 more rows

# Compute Fourier frequencies for a 2D matrix
matrix_input <- matrix(1:9, nrow = 3, ncol = 3)
fourier_frequencies(matrix_input)
#> # A tibble: 9 × 2
#>    dim_2  dim_1
#>    <dbl>  <dbl>
#> 1  0      0    
#> 2  0      0.333
#> 3  0     -0.333
#> 4  0.333  0    
#> 5  0.333  0.333
#> 6  0.333 -0.333
#> 7 -0.333  0    
#> 8 -0.333  0.333
#> 9 -0.333 -0.333

# Compute Fourier frequencies for a vector (1D case)
vector_input <- 1:8
fourier_frequencies(vector_input)
#> # A tibble: 8 × 1
#>    dim_1
#>    <dbl>
#> 1  0    
#> 2  0.125
#> 3  0.25 
#> 4  0.375
#> 5  0.5  
#> 6 -0.375
#> 7 -0.25 
#> 8 -0.125