Perform FFT and IFFT with Tidy Results

Provides functions to compute the Fast Fourier Transform (FFT) and its inverse (IFFT) while maintaining results in a tidy format using tibbles. Supports vectors, time series (ts), and arrays as inputs.

Usage

tidy_fft(x, norm = FALSE)

# Default S3 method
tidy_fft(x, norm = FALSE)

# S3 method for class 'ts'
tidy_fft(x, norm = FALSE)

# S3 method for class 'array'
tidy_fft(x, norm = FALSE)

tidy_ifft(x)

Arguments

x

Input object for which to compute the FFT or IFFT. This can be:

A numeric vector (default method for tidy_fft).

A time series object (ts) for tidy_fft.ts.

A multidimensional numeric array for tidy_fft.array.

A tidy_fft object for tidy_ifft.

norm

Logical. If TRUE, computes normalized coefficients for FFT.

Value

tidy_fft: A tibble containing:

Fourier frequencies (.dim_1, .dim_2, etc.).

FFT values stored in the fx column as complex values.

tidy_ifft: A vector, array, or time series object representing the reconstructed signal.

Details

tidy_fft organizes FFT results into a tidy tibble for downstream analysis.

tidy_ifft ensures that reconstructed signals match the input structure (e.g., arrays, ts).

FFT

The tidy_fft function computes the FFT for different input types:

Default Input (tidy_fft.default): Computes FFT for numeric vectors.

Time Series Input (tidy_fft.ts): Handles FFT for ts objects, scaling frequencies appropriately.

Array Input (tidy_fft.array): Processes multidimensional arrays.

Results are returned as tidy tibbles containing Fourier frequencies and FFT values.

IFFT

The tidy_ifft function reconstructs the original signal from a tidy_fft object. It supports vectors, arrays, and time series inputs. The inverse transform preserves the original structure (e.g., array dimensions or time series attributes).

See also

Examples

tidy_fft(c(1, 0, -1, 0))
#> # A tibble: 4 × 2
#>   .dim_1 fx   
#> *  <dbl> <cpl>
#> 1   0    0+0i 
#> 2   0.25 2+0i 
#> 3   0.5  0+0i 
#> 4  -0.25 2+0i 

tidy_fft(c(1, 0, -1, 0)) |> tidy_ifft()
#> [1]  1  0 -1  0

ts(sin(1:10), frequency = 12) |> tidy_fft()
#> # A tibble: 10 × 2
#>    .dim_1 fx                   
#>  *  <dbl> <cpl>                
#>  1    0    1.4111884+0.0000000i
#>  2    1.2  1.7597704+2.0140018i
#>  3    2.4 -0.0540927-3.7860893i
#>  4    3.6  0.6552682-1.0310219i
#>  5    4.8  0.7536372-0.4010851i
#>  6    6    0.7743552+0.0000000i
#>  7   -4.8  0.7536372+0.4010851i
#>  8   -3.6  0.6552682+1.0310219i
#>  9   -2.4 -0.0540927+3.7860893i
#> 10   -1.2  1.7597704-2.0140018i

array(1:8, dim = c(2, 2, 2)) |> tidy_fft()
#> # A tibble: 8 × 4
#>   .dim_1 .dim_2 .dim_3 fx    
#> *  <dbl>  <dbl>  <dbl> <cpl> 
#> 1    0      0      0    36+0i
#> 2    0.5    0      0    -4+0i
#> 3    0      0.5    0    -8+0i
#> 4    0.5    0.5    0     0+0i
#> 5    0      0      0.5 -16+0i
#> 6    0.5    0      0.5   0+0i
#> 7    0      0.5    0.5   0+0i
#> 8    0.5    0.5    0.5   0+0i