Compute FFT and Return a Tidy Result with Normalized Coefficients

Computes the Fast Fourier Transform (FFT) for various input types, including vectors, time series (ts), and arrays. The FFT values are normalized by the length of the input (or its product of dimensions for arrays). The result is returned as a tidy tibble containing the Fourier frequencies and normalized FFT values.

Usage

tidy_fft(x)

# Default S3 method
tidy_fft(x)

# S3 method for class 'ts'
tidy_fft(x)

# S3 method for class 'array'
tidy_fft(x)

Arguments

x

Input object for which to compute the FFT. This can be:

A numeric vector (default method).

A time series object (ts).

A multidimensional numeric array.

Value

A tibble containing:

Fourier frequencies: Represented by columns dim_1, dim_2, ..., depending on the input dimensions.

Normalized FFT values: Stored in the fx column as complex values, normalized by the input size.

Details

This is a generic function with methods for specific input types:

Default Input (tidy_fft.default): Computes the FFT for a numeric vector and normalizes the result by the vector length.

Time Series Input (tidy_fft.ts): Computes the FFT for a ts object, scaling the frequencies by the time series frequency attribute and normalizing the FFT values.

Array Input (tidy_fft.array): Computes the FFT for a multidimensional array and normalizes the FFT values by the product of the array dimensions.

Each method returns a tidy tibble where the Fourier frequencies (dim_1, dim_2, etc.) are paired with their corresponding normalized FFT values.

See also

Examples

# FFT for a numeric vector
tidy_fft(c(1, 0, -1, 0))
#> # A tibble: 4 × 2
#>   dim_1 fx    
#> * <dbl> <cpl> 
#> 1  0    0.0+0i
#> 2  0.25 0.5+0i
#> 3  0.5  0.0+0i
#> 4 -0.25 0.5+0i

# FFT for a time series
ts_obj <- ts(sin(1:10), frequency = 12)
tidy_fft(ts_obj)
#> # A tibble: 10 × 2
#>    dim_1 fx                     
#>  * <dbl> <cpl>                  
#>  1   0    0.14111884+0.00000000i
#>  2   1.2  0.17597704+0.20140018i
#>  3   2.4 -0.00540927-0.37860893i
#>  4   3.6  0.06552682-0.10310219i
#>  5   4.8  0.07536372-0.04010851i
#>  6   6    0.07743552+0.00000000i
#>  7  -4.8  0.07536372+0.04010851i
#>  8  -3.6  0.06552682+0.10310219i
#>  9  -2.4 -0.00540927+0.37860893i
#> 10  -1.2  0.17597704-0.20140018i

# FFT for a multidimensional array
arr <- array(1:8, dim = c(2, 2, 2))
tidy_fft(arr)
#> # A tibble: 8 × 4
#>   dim_3 dim_2 dim_1 fx     
#> * <dbl> <dbl> <dbl> <cpl>  
#> 1   0     0     0    4.5+0i
#> 2   0     0     0.5 -0.5+0i
#> 3   0     0.5   0   -1.0+0i
#> 4   0     0.5   0.5  0.0+0i
#> 5   0.5   0     0   -2.0+0i
#> 6   0.5   0     0.5  0.0+0i
#> 7   0.5   0.5   0    0.0+0i
#> 8   0.5   0.5   0.5  0.0+0i